# src/swagger/wompi-payouts.yaml
...
{
  "openapi": "3.0.2",
  "info":
    {
      "title": "Payouts API",
      "description": "API público para realizar pagos a terceros través de Wompi.",
      "version": "1.0.0",
    },
  "servers":
    [
      { "url": "https://api.payouts.wompi.co/v1" },
      { "url": "https://api.sandbox.payouts.wompi.co/v1" },
    ],
...
  "paths":
...
      "/payouts":
        {
          "post":
            {
              "tags": ["Payouts"],
              "summary": "Crear una orden de pago.",
              "description": "Crea una nueva orden de pago.",
              "requestBody":
                {
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "oneOf":
                                [
                                  {
                                    "$ref": "#/components/schemas/CreateImmediatePayout",
                                  },
                                  {
                                    "$ref": "#/components/schemas/CreateScheduledPayout",
                                  },
                                  {
                                    "$ref": "#/components/schemas/CreateRecurringPayout",
                                  },
                                ],
                            },
                        },
                    },
                  "required": true,
                },
              "responses":
                {
                  "201":
                    {
                      "description": "OK",
                      "content":
                        {
                          "application/json":
                            {
                              "schema":
                                {
                                  "$ref": "#/components/schemas/PayoutCreateResponse",
                                },
                            },
                        },
                    },
                  "403": { "description": "Forbidden" },
                  "400":
                    {
                      "description": "Error al crear el pago",
                      "content":
                        {
                          "application/json":
                            {
                              "schema":
                                {
                                  "$ref": "#/components/schemas/CreatePayoutError",
                                },
                            },
                        },
                    },
                },
              "security": [{ "x-api-key": [], "user-principal-id": [] }],
            },
          "get":
            {
              "tags": ["Queries"],
              "summary": "Regresa todos los lotes.",
              "description": "Regresa todos los lotes con soporte de paginación.",
              "parameters":
                [
                  {
                    "name": "status",
                    "in": "query",
                    "required": false,
                    "style": "form",
                    "explode": true,
                    "schema":
                      {
                        "type": "array",
                        "items":
                          {
                            "type": "string",
                            "enum":
                              [
                                "PENDING",
                                "REJECTED",
                                "TOTAL_PAYMENT",
                                "PARTIAL_PAYMENT",
                                "PENDING_APPROVAL",
                                "NOT_APPROVED",
                                "AFE_REJECTED",
                                "AFE_ON_HOLD",
                              ],
                          },
                      },
                  },
                  {
                    "name": "fromDate",
                    "in": "query",
                    "required": false,
                    "style": "form",
                    "explode": false,
                    "schema": { "type": "string", "format": "2024-01-01" },
                  },
                  {
                    "name": "toDate",
                    "in": "query",
                    "required": false,
                    "style": "form",
                    "explode": false,
                    "schema": { "type": "string", "format": "2024-01-01" },
                  },
                  {
                    "name": "reference",
                    "in": "query",
                    "required": false,
                    "style": "form",
                    "explode": false,
                    "schema": { "type": "string" },
                  },
                  {
                    "name": "id",
                    "in": "query",
                    "required": false,
                    "style": "form",
                    "explode": false,
                    "schema": { "type": "string" },
                  },
                  {
                    "name": "page",
                    "in": "query",
                    "required": false,
                    "schema": { "type": "integer", "default": 1, "minimum": 1 },
                    "description": "Número de la página que se quiere consultar.",
                  },
                  {
                    "name": "limit",
                    "in": "query",
                    "required": false,
                    "schema":
                      { "type": "integer", "default": 10, "minimum": 1 },
                    "description": "Cantidad de elementos por página.",
                  },
                ],
              "responses":
                {
                  "200":
                    {
                      "description": "OK",
                      "content":
                        {
                          "application/json":
                            {
                              "schema":
                                {
                                  "type": "object",
                                  "properties":
                                    {
                                      "status":
                                        { "type": "integer", "example": 200 },
                                      "meta":
                                        {
                                          "type": "object",
                                          "properties":
                                            {
                                              "trace_id":
                                                {
                                                  "type": "string",
                                                  "format": "uuid",
                                                  "example": "2796d870-21e5-11f0-805b-4b055319fe18",
                                                },
                                            },
                                          "required": ["trace_id"],
                                        },
                                      "code":
                                        { "type": "string", "example": "OK" },
                                      "message":
                                        {
                                          "type": "string",
                                          "example": "Solicitud ejecutada correctamente.",
                                        },
                                      "data":
                                        {
                                          "type": "object",
                                          "properties":
                                            {
                                              "page":
                                                {
                                                  "type": "integer",
                                                  "example": 1,
                                                },
                                              "limit":
                                                {
                                                  "type": "integer",
                                                  "example": 10,
                                                },
                                              "total":
                                                {
                                                  "type": "integer",
                                                  "example": 53,
                                                },
                                              "pages":
                                                {
                                                  "type": "integer",
                                                  "example": 6,
                                                },
                                              "records":
                                                {
                                                  "type": "array",
                                                  "items":
                                                    {
                                                      "$ref": "#/components/schemas/PayoutResponse",
                                                    },
                                                },
                                            },
                                        },
                                    },
                                },
                            },
                        },
                    },
                },
              "security": [{ "x-api-key": [], "user-principal-id": [] }],
            },
        },
...
      "/banks":
        {
          "get":
            {
              "tags": ["Resources"],
              "summary": "Regresa la lista de bancos.",
              "description": "Regresa la lista de bancos disponibles para realizar pagos.",
              "responses":
                {
                  "200":
                    {
                      "description": "OK",
                      "content":
                        {
                          "application/json":
                            {
                              "schema":
                                {
                                  "type": "array",
                                  "items":
                                    {
                                      "$ref": "#/components/schemas/inline_response_200",
                                    },
                                },
                            },
                        },
                    },
                },
              "security": [{ "x-api-key": [], "user-principal-id": [] }],
            },
        },
      "/accounts":
        {
          "get":
            {
              "tags": ["Resources"],
              "summary": "Regresa la lista de cuentas.",
              "description": "Regresa la lista de cuentas disponibles para dispersar los pagos.",
              "parameters":
                [
                  {
                    "name": "bankCodes",
                    "in": "query",
                    "required": false,
                    "style": "form",
                    "explode": false,
                    "schema":
                      { "type": "array", "items": { "type": "string" } },
                  },

                  {
                    "name": "status",
                    "in": "query",
                    "required": false,
                    "style": "form",
                    "explode": false,
                    "schema":
                      {
                        "type": "array",
                        "items":
                          {
                            "type": "string",
                            "enum": ["IN_REVIEW", "ACTIVE", "INACTIVE"],
                          },
                      },
                  },
                ],
              "responses":
                {
                  "200":
                    {
                      "description": "OK",
                      "content":
                        {
                          "application/json":
                            {
                              "schema":
                                {
                                  "type": "object",
                                  "properties":
                                    {
                                      "status":
                                        { "type": "integer", "example": 200 },
                                      "meta":
                                        {
                                          "type": "object",
                                          "properties":
                                            {
                                              "trace_id":
                                                {
                                                  "type": "string",
                                                  "format": "uuid",
                                                  "example": "2796d870-21e5-11f0-805b-4b055319fe18",
                                                },
                                            },
                                          "required": ["trace_id"],
                                        },
                                      "code":
                                        { "type": "string", "example": "OK" },
                                      "message":
                                        {
                                          "type": "string",
                                          "example": "Solicitud ejecutada correctamente.",
                                        },
                                      "data":
                                        {
                                          "type": "array",
                                          "items":
                                            {
                                              "$ref": "#/components/schemas/inline_response_200_1",
                                            },
                                        },
                                    },
                                },
                            },
                        },
                    },
                },
              "security": [{ "x-api-key": [], "user-principal-id": [] }],
            },
        },
      "/limits":
        {
          "get":
            {
              "tags": ["Resources"],
              "summary": "Regresa los límites y transacciones.",
              "description": "Regresa los límites diarios, incluyendo el tope diario, disponible y consumido. En caso de tener un plan por suscripción, también regresará las transacciones asociadas al plan.",
              "responses":
                {
                  "200":
                    {
                      "description": "OK",
                      "content":
                        {
                          "application/json":
                            {
                              "schema":
                                {
                                  "$ref": "#/components/schemas/LimitsResponse",
                                },
                            },
                        },
                    },
                },
              "security": [{ "x-api-key": [], "user-principal-id": [] }],
            },
        },
...
    },
  "components":
...
      "schemas":
...
          "CreateImmediatePayout":
            {
              "required":
                ["accountId", "paymentType", "reference", "transactions"],
              "type": "object",
              "properties":
                {
                  "reference": { "type": "string", "example": "reference" },
                  "accountId": { "type": "string", "example": "account-id" },
                  "paymentType":
                    {
                      "type": "string",
                      "example": "PAYROLL",
                      "enum": ["PAYROLL", "PROVIDERS", "OTHER"],
                    },
                  "transactionStatus":
                    {
                      "type": "string",
                      "description": "Solo en sandbox. Emular el estado final de la transacción.",
                      "enum": ["APPROVED", "FAILED"],
                      "example": "FAILED // solo Sandbox",
                    },
                  "transactions":
                    {
                      "type": "array",
                      "items":
                        { "$ref": "#/components/schemas/CreateTransaction" },
                    },
                },
            },
          "CreateScheduledPayout":
            {
              "allOf":
                [
                  { "$ref": "#/components/schemas/CreateImmediatePayout" },
                  {
                    "type": "object",
                    "required": ["dispersionDatetime"],
                    "properties":
                      {
                        "dispersionDatetime":
                          {
                            "type": "string",
                            "format": "date-time",
                            "example": "2025-05-05T12:30",
                          },
                      },
                  },
                ],
            },
          "CreateRecurringPayout":
            {
              "allOf":
                [
                  { "$ref": "#/components/schemas/CreateScheduledPayout" },
                  {
                    "type": "object",
                    "required": ["recurring"],
                    "properties":
                      {
                        "recurring":
                          {
                            "type": "object",
                            "required": ["interval", "months"],
                            "properties":
                              {
                                "interval":
                                  {
                                    "type": "string",
                                    "enum": ["biweek", "month"],
                                  },
                                "months":
                                  { "type": "integer", "enum": [3, 6, 12] },
                                "description":
                                  {
                                    "type": "string",
                                    "example": "Nómina colaboradores",
                                  },
                              },
                          },
                      },
                  },
                ],
            },
          "PayoutCreateResponse":
            {
              "type": "object",
              "properties":
                {
                  "status": { "type": "integer", "example": 201 },
                  "meta":
                    {
                      "$ref": "#/components/schemas/PayoutCreateResponse_meta",
                    },
                  "code": { "type": "string", "example": "OK" },
                  "message":
                    {
                      "type": "string",
                      "example": "Solicitud ejecutada correctamente.",
                    },
                  "data":
                    {
                      "$ref": "#/components/schemas/PayoutCreateResponse_data",
                    },
                },
            },
          "CreateTransaction":
            {
              "required":
                [
                  "accountNumber",
                  "accountType",
                  "amount",
                  "bankId",
                  "legalId",
                  "legalIdType",
                  "name",
                  "personType",
                ],
              "type": "object",
              "properties":
                {
                  "legalIdType":
                    {
                      "type": "string",
                      "example": "CC",
                      "enum": ["CC", "NIT", "PP", "CE", "TI", "DNI"],
                    },
                  "legalId": { "type": "string", "example": "1234567890" },
                  "bankId": { "type": "string", "example": "bankId" },
                  "accountType":
                    {
                      "type": "string",
                      "example": "AHORROS",
                      "enum": ["AHORROS", "CORRIENTE", "DEPOSITO_ELECTRONICO"],
                    },
                  "accountNumber":
                    { "type": "string", "example": "1234567890" },
                  "name": { "type": "string", "example": "John Doe" },
                  "amount": { "type": "integer", "example": 1000 },
                  "personType":
                    {
                      "type": "string",
                      "example": "NATURAL",
                      "enum": ["NATURAL", "JURIDICA"],
                    },
                  "description": { "type": "string", "example": "description" },
                  "phone": { "type": "string", "example": "1234567890" },
                  "email": { "type": "string", "example": "email@example.com" },
                  "reference": { "type": "string", "example": "ref1234567890" },
                },
            },
          "CreatePayoutError":
            {
              "type": "object",
              "properties":
                {
                  "code": { "type": "string", "example": "EXC_003" },
                  "message":
                    {
                      "type": "string",
                      "example": "Se ha alcanzado el límite de lotes diarios.",
                    },
                  "type": { "type": "string", "example": "Business" },
                  "meta":
                    {
                      "$ref": "#/components/schemas/PayoutCreateResponse_meta",
                    },
                },
            },
...
          "inline_response_200":
            {
              "type": "object",
              "properties":
                {
                  "id":
                    {
                      "type": "string",
                      "example": "00000000-0000-0000-0000-000000000000",
                    },
                  "name": { "type": "string", "example": "Bancolombia" },
                  "code": { "type": "string", "example": "BANCOLOMBIA" },
                  "achCode": { "type": "integer", "example": "1507" },
                  "swiftCode": { "type": "string", "example": null },
                  "allowedForOrigin": { "type": "boolean", "example": true },
                  "isElectronicDeposit": { "type": "boolean", "example": true },
                  "createdAt":
                    { "type": "string", "example": "2024-01-01T23:23:23.322Z" },
                  "updatedAt":
                    { "type": "string", "example": "2024-02-01T23:23:23.322Z" },
                },
            },
...
          "inline_response_200_1":
            {
              "type": "object",
              "properties":
                {
                  "id":
                    {
                      "type": "string",
                      "example": "00000000-0000-0000-0000-000000000000",
                    },
                  "balanceInCents": { "type": "integer", "example": 1000 },
                  "number": { "type": "string", "example": "****7890" },
                  "status": { "type": "string", "example": "ACTIVE" },
                  "accountType":
                    {
                      "type": "string",
                      "example": "AHORROS",
                      "enum": ["AHORROS", "CORRIENTE", "DEPOSITO_ELECTRONICO"],
                    },
                  "bank": { "$ref": "#/components/schemas/accounts_bank" },
                  "updatedAt":
                    { "type": "string", "example": "2024-01-01T23:23:23.322Z" },
                },
            },
          "PayoutCreateResponse_meta":
            {
              "type": "object",
              "properties":
                {
                  "trace_id":
                    {
                      "type": "string",
                      "example": "00000000-0000-0000-0000-000000000000",
                    },
                },
            },
          "PayoutCreateResponse_data":
            {
              "type": "object",
              "properties":
                {
                  "payoutId":
                    {
                      "type": "string",
                      "example": "00000000-0000-0000-0000-000000000000",
                    },
                  "transactions": { "type": "integer", "example": 3 },
                  "success": { "type": "integer", "example": 1 },
                  "failed": { "type": "integer", "example": 2 },
                },
            },
          "PayoutResponse":
            {
              "type": "object",
              "properties":
                {
                  "id":
                    {
                      "type": "string",
                      "format": "uuid",
                      "example": "2b184a70-1947-11f0-b2d8-1337d22b135c",
                    },
                  "reference": { "type": "string", "example": "reference" },
                  "type": { "type": "string", "example": "INMEDIATE" },
                  "amountInCents": { "type": "integer", "example": 300 },
                  "status": { "type": "string", "example": "REJECTED" },
                  "payerInfo":
                    {
                      "type": "object",
                      "properties":
                        {
                          "name":
                            {
                              "type": "string",
                              "example": "PruebasAPIsPayout",
                            },
                          "email":
                            {
                              "type": "string",
                              "format": "email",
                              "example": "luisa.aguilar+pruebasapis@wompi.com",
                            },
                          "phone":
                            { "type": "string", "example": "3125215252" },
                          "legalId":
                            { "type": "string", "example": "123456789" },
                          "personType":
                            { "type": "string", "example": "NATURAL" },
                          "legalIdType": { "type": "string", "example": "CC" },
                        },
                      "required":
                        [
                          "name",
                          "email",
                          "phone",
                          "legalId",
                          "personType",
                          "legalIdType",
                        ],
                    },
                  "statusMessage": { "type": "string", "example": "" },
                  "totalTransactions": { "type": "integer", "example": 3 },
                  "dispersionDatetime":
                    {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-04-14T20:43:00.000Z",
                    },
                  "paymentType": { "type": "string", "example": "PROVIDERS" },
                  "createdAt":
                    {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-04-14T20:43:00.000Z",
                    },
                },
              "required":
                [
                  "id",
                  "reference",
                  "type",
                  "amountInCents",
                  "status",
                  "payerInfo",
                  "statusMessage",
                  "totalTransactions",
                  "dispersionDatetime",
                  "paymentType",
                  "createdAt",
                ],
            },
...
          "LimitsResponse":
            {
              "type": "object",
              "properties":
                {
                  "numberOfTransactionsConsumed":
                    { "type": "integer", "example": 22 },
                  "limits":
                    {
                      "type": "object",
                      "properties":
                        {
                          "dailyLimit":
                            { "type": "integer", "example": 1500000000 },
                          "dailyAvailable":
                            { "type": "integer", "example": 1499800000 },
                          "dailyConsumed":
                            { "type": "integer", "example": 200000 },
                        },
                    },
                },
            },
...
        },
      "securitySchemes":
        {
          "x-api-key":
            { "type": "apiKey", "name": "x-api-key", "in": "header" },
          "user-principal-id":
            { "type": "apiKey", "name": "user-principal-id", "in": "header" },
        },
    },
}